// src/app/models/deck.ts
export class Deck {
    ID: Number;
    Player_id: Number;
    Deck_id: Number;
    Hero_id: Number;
    Name: String;
    Card_1_id: Number;
    Card_2_id: Number;
    Card_3_id: Number;
    Card_4_id: Number;
    Card_5_id: Number;
    Card_6_id: Number;
    Card_7_id: Number;
    Card_8_id: Number;
    Card_9_id: Number;
    Card_10_id: Number;
    Card_11_id: Number;
    Card_12_id: Number;
    Card_13_id: Number;
    Card_14_id: Number;
    Card_15_id: Number;
    Card_16_id: Number;
    Card_17_id: Number;
    Card_18_id: Number;
    Card_19_id: Number;
    Card_20_id: Number;

    constructor(ID, Player_id, Deck_id, Hero_id, Name, Card_1_id, Card_2_id, Card_3_id, Card_4_id, Card_5_id, Card_6_id, Card_7_id,
    Card_8_id, Card_9_id, Card_10_id, Card_11_id, Card_12_id, Card_13_id, Card_14_id, Card_15_id, Card_16_id, Card_17_id,
    Card_18_id, Card_19_id, Card_20_id){
        this.ID = ID;
        this.Player_id = Player_id;
        this.Deck_id = Deck_id;
        this.Hero_id = Hero_id;
        this.Name = Name;
        this.Card_1_id = Card_1_id;
        this.Card_2_id = Card_2_id;
        this.Card_3_id = Card_3_id;
        this.Card_4_id = Card_4_id;
        this.Card_5_id = Card_5_id;
        this.Card_6_id = Card_6_id;
        this.Card_7_id = Card_7_id;
        this.Card_8_id = Card_8_id;
        this.Card_9_id = Card_9_id;
        this.Card_10_id = Card_10_id;
        this.Card_11_id = Card_11_id;
        this.Card_12_id = Card_12_id;
        this.Card_13_id = Card_13_id;
        this.Card_14_id = Card_14_id;
        this.Card_15_id = Card_15_id;
        this.Card_16_id = Card_16_id;
        this.Card_17_id = Card_17_id;
        this.Card_18_id = Card_18_id;
        this.Card_19_id = Card_19_id;
        this.Card_20_id = Card_20_id;
    }
}